Change account nonce to uint64_t

+++ test/state/account.hpp
@@ -20,7 +20,7 @@ struct StorageValue
 struct Account
 {
     /// The account nonce.
-    int nonce = 0;
+    uint64_t nonce = 0;
 
     /// The account code.
     evmc::bytes code;
